Implementation of the BashDoc monoid
The BashDoc monoid assist during the pretty printing of Bash scripts. It
tracks the following things:
- The first line of the document in question
- Any lines following the first line. We refer to those as the rest.
- All here documents which are pending

When two BashDocs are appended the here documents of the first one are
inserted between the first lines and the rest of the first monoid and
the rest of the second one. This ensures that commands in the first part
of e.g. a pipeline can still access their here documents while the
following parts, for example the do-block of a loop, appear below them.
This results in a valid Bash script.